Add basic click listener to tale adapter

+++ app/src/main/java/com/ojomono/ionce/ui/tales/Tale.kt
@@ -1,3 +1,3 @@
 package com.ojomono.ionce.ui.tales
 
-data class Tale(val id: Int, val title: String)
+data class Tale(val id: Long, val title: String)
